pyproject.toml
[project]
name = "clip-local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "accelerate>=0.33.0",
    "fastapi>=0.129.0",
    "huggingface-hub>=1.4.1",
    "modal>=1.3.4",
    "pillow>=12.1.1",
    "pypdf>=4.3.1",
    "python-multipart>=0.0.22",
    "sentencepiece>=0.2.0",
    "torch>=2.10.0",
    "transformers @ git+https://github.com/huggingface/transformers.git",
    "uvicorn[standard]>=0.41.0",
]
